+++ albumentations/augmentations/geometric/transforms.py
@@ -50,6 +50,7 @@ __all__ = [
     "PadIfNeeded",
     "D4",
     "GridElasticDeform",
+    "RandomGridShuffle",
 ]
 
 
@@ -2028,3 +2029,108 @@ class GridElasticDeform(DualTransform):
 
     def get_transform_init_args_names(self) -> tuple[str, ...]:
         return "num_grid_xy", "magnitude", "interpolation", "mask_interpolation"
+
+
+class RandomGridShuffle(DualTransform):
+    """Randomly shuffles the grid's cells on an image, mask, or keypoints,
+    effectively rearranging patches within the image.
+    This transformation divides the image into a grid and then permutes these grid cells based on a random mapping.
+
+    Args:
+        grid (tuple[int, int]): Size of the grid for splitting the image into cells. Each cell is shuffled randomly.
+            For example, (3, 3) will divide the image into a 3x3 grid, resulting in 9 cells to be shuffled.
+            Default: (3, 3)
+        p (float): Probability that the transform will be applied. Should be in the range [0, 1].
+            Default: 0.5
+
+    Targets:
+        image, mask, keypoints
+
+    Image types:
+        uint8, float32
+
+    Note:
+        - This transform maintains consistency across all targets. If applied to an image and its corresponding
+          mask or keypoints, the same shuffling will be applied to all.
+        - The number of cells in the grid should be at least 2 (i.e., grid should be at least (1, 2), (2, 1), or (2, 2))
+          for the transform to have any effect.
+        - Keypoints are moved along with their corresponding grid cell.
+        - This transform could be useful when only micro features are important for the model, and memorizing
+          the global structure could be harmful. For example:
+          - Identifying the type of cell phone used to take a picture based on micro artifacts generated by
+            phone post-processing algorithms, rather than the semantic features of the photo.
+            See more at https://ieeexplore.ieee.org/abstract/document/8622031
+          - Identifying stress, glucose, hydration levels based on skin images.
+
+    Mathematical Formulation:
+        1. The image is divided into a grid of size (m, n) as specified by the 'grid' parameter.
+        2. A random permutation P of integers from 0 to (m*n - 1) is generated.
+        3. Each cell in the grid is assigned a number from 0 to (m*n - 1) in row-major order.
+        4. The cells are then rearranged according to the permutation P.
+
+    Example:
+        >>> import numpy as np
+        >>> import albumentations as A
+        >>> image = np.array([
+        ...     [1, 1, 1, 2, 2, 2],
+        ...     [1, 1, 1, 2, 2, 2],
+        ...     [1, 1, 1, 2, 2, 2],
+        ...     [3, 3, 3, 4, 4, 4],
+        ...     [3, 3, 3, 4, 4, 4],
+        ...     [3, 3, 3, 4, 4, 4]
+        ... ])
+        >>> transform = A.RandomGridShuffle(grid=(2, 2), p=1.0)
+        >>> result = transform(image=image)
+        >>> transformed_image = result['image']
+        # The resulting image might look like this (one possible outcome):
+        # [[4, 4, 4, 2, 2, 2],
+        #  [4, 4, 4, 2, 2, 2],
+        #  [4, 4, 4, 2, 2, 2],
+        #  [3, 3, 3, 1, 1, 1],
+        #  [3, 3, 3, 1, 1, 1],
+        #  [3, 3, 3, 1, 1, 1]]
+
+    """
+
+    class InitSchema(BaseTransformInitSchema):
+        grid: Annotated[tuple[int, int], AfterValidator(check_1plus)]
+
+    _targets = (Targets.IMAGE, Targets.MASK, Targets.KEYPOINTS)
+
+    def __init__(self, grid: tuple[int, int] = (3, 3), p: float = 0.5, always_apply: bool | None = None):
+        super().__init__(p=p, always_apply=always_apply)
+        self.grid = grid
+
+    def apply(self, img: np.ndarray, tiles: np.ndarray, mapping: list[int], **params: Any) -> np.ndarray:
+        return fgeometric.swap_tiles_on_image(img, tiles, mapping)
+
+    def apply_to_bboxes(self, bboxes: np.ndarray, tiles: np.ndarray, mapping: np.ndarray, **params: Any) -> np.ndarray:
+        image_shape = params["shape"][:2]
+        bboxes_denorm = denormalize_bboxes(bboxes, image_shape)
+        bboxes_returned = fgeometric.bboxes_grid_shuffle(bboxes_denorm, tiles, mapping, image_shape)
+        return normalize_bboxes(bboxes_returned, image_shape)
+
+    def apply_to_keypoints(
+        self,
+        keypoints: np.ndarray,
+        tiles: np.ndarray,
+        mapping: np.ndarray,
+        **params: Any,
+    ) -> np.ndarray:
+        return fgeometric.swap_tiles_on_keypoints(keypoints, tiles, mapping)
+
+    def get_params_dependent_on_data(self, params: dict[str, Any], data: dict[str, Any]) -> dict[str, np.ndarray]:
+        image_shape = params["shape"][:2]
+
+        original_tiles = fgeometric.split_uniform_grid(
+            image_shape,
+            self.grid,
+            self.random_generator,
+        )
+        shape_groups = fgeometric.create_shape_groups(original_tiles)
+        mapping = fgeometric.shuffle_tiles_within_shape_groups(shape_groups, self.random_generator)
+
+        return {"tiles": original_tiles, "mapping": mapping}
+
+    def get_transform_init_args_names(self) -> tuple[str, ...]:
+        return ("grid",)
